fix bugs in service and incidentItems

+++ WIL/SystemLogic/Truck.cs
@@ -60,7 +60,7 @@ namespace SystemLogic
         public int EngineSize { get; set; }
         public int ServiceInterval { get; set; }
         public int MaxWeight { get; set; }
-        public float LitersPerHundy { get; set; }
+        public int LitersPerHundy { get; set; }
         public int MaxVol { get; set; }
 
         //default constructor
@@ -71,7 +71,7 @@ namespace SystemLogic
 
         //constructor will all paramaters
         public TruckType(int id, string type, string man,
-            int engineSize, int serviceInterval, int maxWeight, float litersPerHundy, int maxVol)
+            int engineSize, int serviceInterval, int maxWeight, int litersPerHundy, int maxVol)
         {
             ID = id;
             this.Type = type;
@@ -85,7 +85,7 @@ namespace SystemLogic
 
         //constructor without ID
         public TruckType(string type, string man,
-            int engineSize, int serviceInterval, int maxWeight, float litersPerHundy, int maxVol)
+            int engineSize, int serviceInterval, int maxWeight, int litersPerHundy, int maxVol)
         {
             this.Type = type;
             Manufacturor = man;
